Move AWS-specific modules into provider-scoped packages
Group enrichment modules whose extraction logic is AWS-specific under
modules/aws/ (Networking, Serverless) and modules/ccf/aws/ (Storage,
Accelerators), so equivalent classes for other cloud providers can sit
alongside without name clashes. Drop the deprecated ccf flat-rate
Networking module that was already superseded by aws.Networking.
Default AWS config and module documentation updated to match.
