+++ src/main/java/com/digitalpebble/spruce/modules/aws/Serverless.java
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalpebble.spruce.modules;
+package com.digitalpebble.spruce.modules.aws;
 
 import com.digitalpebble.spruce.Column;
 import com.digitalpebble.spruce.EnrichmentModule;
