+++ src/main/java/com/digitalpebble/spruce/modules/aws/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Enrichment modules whose extraction logic is specific to Amazon Web Services
+ * (e.g. AWS CUR usage types, service codes). Sibling packages will hold the
+ * equivalents for other cloud providers.
+ */
+package com.digitalpebble.spruce.modules.aws;
