--- src/main/java/com/digitalpebble/spruce/modules/ccf/Networking.java
@@ -1,73 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-package com.digitalpebble.spruce.modules.ccf;
-
-import com.digitalpebble.spruce.Column;
-import com.digitalpebble.spruce.EnrichmentModule;
-import com.digitalpebble.spruce.Utils;
-import org.apache.spark.sql.Row;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.Map;
-
-import static com.digitalpebble.spruce.CURColumn.*;
-import static com.digitalpebble.spruce.SpruceColumn.ENERGY_USED;
-
-/**
- * Provides an estimate of energy used for networking in and out of data centres.
- * Applies a flat coefficient per Gb
- *
- * @see <a href="https://www.cloudcarbonfootprint.org/docs/methodology#networking">CCF methodology</a>
- * @see <a href="https://github.com/cloud-carbon-footprint/cloud-carbon-footprint/blob/9f2cf436e5ad020830977e52c3b0a1719d20a8b9/packages/aws/src/lib/CostAndUsageTypes.ts#L25">resource file</a>
- *
- * @deprecated com.digitalpebble.spruce.modules.Networking is a more accurate replacement
- **/
-public class Networking implements EnrichmentModule {
-
-    private static final Logger log = LoggerFactory.getLogger(Networking.class);
-
-    // estimated kWh/Gb
-    double network_coefficient = 0.001;
-
-    @Override
-    public void init(Map<String, Object> params) {
-        Double coef = (Double) params.get("network_coefficient");
-        if (coef != null) {
-            network_coefficient = coef;
-        }
-        log.info("network_coefficient: {}", network_coefficient);
-    }
-
-    @Override
-    public Column[] columnsNeeded() {
-        return new Column[]{PRODUCT_SERVICE_CODE, PRODUCT, USAGE_AMOUNT};
-    }
-
-    @Override
-    public Column[] columnsAdded() {
-        return new Column[]{ENERGY_USED};
-    }
-
-    @Override
-    public void enrich(Row row, Map<Column, Object> enrichedValues) {
-        String service_code = PRODUCT_SERVICE_CODE.getString(row);
-        if (service_code == null || !service_code.equals("AWSDataTransfer")) {
-            return;
-        }
-        //  apply only to rows corresponding to networking in or out of a region
-        String transfer_type = Utils.getStringFromProductMap(row, "transfer_type", "");
-
-        if (!transfer_type.startsWith("InterRegion")) {
-            return;
-        }
-
-        // TODO consider extending to AWS Outbound and Inbound
-
-        // get the amount of data transferred
-        double amount_gb = USAGE_AMOUNT.getDouble(row);
-        double energy_gb = amount_gb * network_coefficient;
-
-        enrichedValues.put(ENERGY_USED, energy_gb);
-    }
-}
