+++ src/main/java/com/digitalpebble/spruce/modules/ccf/aws/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * AWS-specific implementations of the Cloud Carbon Footprint enrichment
+ * modules. Sibling packages will hold the equivalents for other cloud providers.
+ *
+ * @see <a href="https://www.cloudcarbonfootprint.org/docs/methodology">CCF methodology</a>
+ */
+package com.digitalpebble.spruce.modules.ccf.aws;
