+++ src/main/resources/default-config-aws.json
@@ -4,14 +4,14 @@
       "className": "com.digitalpebble.spruce.modules.RegionExtraction"
     },
     {
-      "className": "com.digitalpebble.spruce.modules.ccf.Storage",
+      "className": "com.digitalpebble.spruce.modules.ccf.aws.Storage",
       "config": {
         "hdd_coefficient_tb_h": 0.65,
         "ssd_coefficient_tb_h": 1.2
       }
     },
     {
-      "className": "com.digitalpebble.spruce.modules.Networking",
+      "className": "com.digitalpebble.spruce.modules.aws.Networking",
       "config": {
         "network_coefficients_kwh_gb": {
           "intra": 0.001,
@@ -24,7 +24,7 @@
       "className": "com.digitalpebble.spruce.modules.boavizta.BoaviztAPIstatic"
     },
     {
-      "className": "com.digitalpebble.spruce.modules.Serverless",
+      "className": "com.digitalpebble.spruce.modules.aws.Serverless",
       "config": {
         "memory_coefficient_kwh": 0.0000598,
         "x86_cpu_coefficient_kwh": 0.0088121875,
@@ -32,7 +32,7 @@
       }
     },
     {
-      "className": "com.digitalpebble.spruce.modules.ccf.Accelerators",
+      "className": "com.digitalpebble.spruce.modules.ccf.aws.Accelerators",
       "config": {
         "gpu_utilisation_percent": 50
       }
